moved api handlers to api subdirectory
