+++ handlers/api/GetObject.go
@@ -1,4 +1,4 @@
-package handlers
+package api
 
 import (
 	"github.com/gin-gonic/gin"
