+++ handlers/api/HeadBucket.go
@@ -1,4 +1,4 @@
-package handlers
+package api
 
 import (
 	"github.com/gin-gonic/gin"
